+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file. This change
 
 ## [1-alpha4-SNAPSHOT]
 - basic change detection (#4)
+- special value representations are defined in an ordered fashion
 
 ## [1-alpha3]
 - shifting to shapshot-based workflow